--- a/components/support/cli/src/lib.rs
+++ b/components/support/cli/src/lib.rs
@@ -7,3 +7,23 @@
 
 pub mod fxa_creds;
 pub mod prompt;
+
+pub use env_logger;
+
+pub fn init_logging_with(s: &str) {
+    let noisy = "tokio_threadpool=warn,tokio_reactor=warn,tokio_core=warn,tokio=warn,hyper=warn,want=warn,mio=warn,reqwest=warn";
+    let spec = format!("{},{}", s, noisy);
+    env_logger::init_from_env(env_logger::Env::default().filter_or("RUST_LOG", spec));
+}
+
+pub fn init_trace_logging() {
+    init_logging_with("trace")
+}
+
+pub fn init_logging() {
+    init_logging_with(if cfg!(debug_assertions) {
+        "debug"
+    } else {
+        "info"
+    })
+}