--- a/components/support/sync15-traits/Cargo.toml
+++ b/components/support/sync15-traits/Cargo.toml
@@ -15,6 +15,4 @@ serde_json = "1"
 log = "0.4"
 ffi-support = "0.4"
 url = "2.1"
-thiserror = "1.0"
 anyhow = "1.0"
-interrupt-support = { path = "../interrupt" }