--- a/examples/fxa-client/src/migration.rs
+++ b/examples/fxa-client/src/migration.rs
@@ -1,3 +1,7 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
 use cli_support::prompt::prompt_string;
 use fxa_client::{Config, FirefoxAccount};
 use std::{thread, time};