--- a/testing/separated/places-tests/src/tests.rs
+++ b/testing/separated/places-tests/src/tests.rs
@@ -2,6 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
+mod check_coop_tx;
 mod fennec_bookmarks;
 mod fennec_history;
 mod ios_bookmarks;