… Cargo.lock
… docs/test-faster.md
… examples/cli-support/Cargo.toml
… examples/cli-support/src/fxa_creds.rs
… examples/cli-support/src/lib.rs
… examples/cli-support/src/prompt.rs
… examples/fxa-client/src/devices-api.rs
… examples/places-autocomplete/Cargo.toml
… examples/places-utils/Cargo.toml
… examples/push-livetest/Cargo.toml
… examples/push-livetest/src/livetest.rs
… examples/sync-pass/Cargo.toml
… examples/tabs-sync/Cargo.toml
… megazords/ios/DEPENDENCIES.md
… testing/separated/README.md
… testing/separated/ffi-support-extra-test/Cargo.toml
… testing/separated/places-bench/Cargo.toml
… testing/separated/places-bench/fixtures/README.md
… testing/separated/places-bench/fixtures/dummy_urls.json
… testing/separated/places-tests/Cargo.toml
… testing/separated/places-tests/src/fennec_bookmarks.rs
… testing/separated/places-tests/src/fennec_bookmarks_schema.sql
… testing/separated/places-tests/src/fennec_history.rs
… testing/separated/places-tests/src/fennec_history_schema.sql
… testing/separated/places-tests/src/ios_bookmarks.rs
… testing/separated/places-tests/src/ios_schema.sql