I spent part of the weekend on a project to fix something that *really* bugged
me about our dev dependencies. specifically, as we get more/better tests and
examples, out build times get worse and IDEs get slower.

I spent this evening finishing that up (took longer than I expected, okay?)

## Quick overview

Most of what I did is the following:
    - Moved all examples to `/examples`
    - Moved benchmarks and slow-building tests to `/testing/separate`
    - Cleaned up some dep usage, fix cases where rebuild triggered for no reason.
    - ...

In between all that, a buncha research on the problem, and wrote a guideline for my findings and when to do these things, why some of the specific choices I made IMO are a good idea, ...

It's late, I haven't yet edited it, but here, test-faster.md is the best overview: https://github.com/mozilla/application-services/pull/3349/files#diff-b1d75bab56cc23929b4b20a7f920ce99.

TLDR for when to do this is:
- always for benchmarks,
- probably always for examples since it's easy,
- when warranted for tests.
    - I'd have liked to do it for the crazy fxa http+struct mocking tests but it's just not plausible. As a result, FxA saw the least benefit here, although in practice it should get benefits from both everything it depends on building faster, and from it triggering way fewer rebuilds of other crates.

## Goals

On for the defense of this major change. As background, my goals were:

- Reducing the minimum time needed to build crates, either by not building stuff we don't need, playing more nicely with cargo features, or making sccache work better.

    - This was a definite success. Some remaining avenues are open, but it went better than hped.

- Reducing the amount rebuilt when changing unrelated crates. For example, currently if `fxa-client` changes, we need to rebuild: `places`, `tabs`, and `logins` completely because of the way our code is organized and not because of a real dependency.

    - This was a definite success

- Reducing the time spent waiting on IDE results.

    - Never found a great way to measure this with any degree of reliability. It feels faster?

## Outcome

Some numbers for single-change-and-rebuild for a few crates. For repeatability:

- `$ cargo test -p $crate`
- Manually change lib.rs in a superficial way
- `$ cargo test -p $crate
- Write down $NUM in `Finished test [unoptimized + debuginfo] target(s) in $NUM`
- All done without sccache, on a Macbook Pro 2018 2.9GHz Intel Core i9 with 12 cores (6 cpu 12 hyper)

1. For logins:
    - before 7.69s
    - after 4.22s
    - 45% speedup

2. For webext-storage:
    - before 4.70s
    - after 2.61s
    - 44% speedup

3. For tabs:
    - before => 6.21s,
    - after => 1.49s,
    - 76% speedup

4. For places:
    - before 13.49s
    - after 7.24s
    - 46% speedup
    - you might point out that the same set of tests aren't running for places --
      this is fair, that moved to a separate crate, and if i time a build
      of the separate crate 10.11s, or 25% speedup.

5. For sync15:
    - before 5.14s
    - after 4.11s
    - 20% speedup

6. ...
7. The only "loser" is fxa-client:
    - Before 13.51s, after between 10.44s and 13.62s. Not very stable.
    - That said.... we know the reason for this.

    - We aren't able to separate heavyweight test dependencies (mockito or mockiato)

    - Most of the time here is spent in a proc macros, a known issue for fxa-client, esp. http_client.rs

    - On the bright side:
        - If fxa-client's dependencies have to rebuild, they all build much faster
        - When fxa-client is modified or otherwise needs rebuilding, logins, places, and tabs no longer require rebuilding.

(Aside: You might also say that moving the examples out is cheating too, but this isn't true -- previously we'd build everything needed for the examples just to be able to run the tests, but it wasn't building examples/benchmarks every time, just their deps (pointlessly), cargo can't tell the difference between an dev-dep an example needs and one a test needs. So I disagree)

Over all build time change:

- Time for `cargo clean; cargo build --test;` without sccache
    - Before: 127.4s (2m 7.4s)
    - After: 116.8s (1m 56.8s)
    - ~8% speedup

There was no reason really to expect this last benchmark to get better -- I was optimizing for other things, but 8% is nothing to sneeze at by any means. I would *guess* the reason it did is mostly because of increased build parallelism. But anyway, I listed the goals up front, this wasn't one of them, so I'm thrilled about that.

Related to full rebuild, this is work is essentially compatible with https://github.com/mozilla/application-services/pull/3329 which should have a more substantial impact for wall time, but admittedly only percieved wall time (it does this by skipping building the megazords and systest for builds that don't request them, among some other things).